Add a default exception handler so that uncaught exceptions are sent to Log4J. Exit the process when OutOfMemoryError occurs.
