+++ teku/src/main/java/tech/pegasys/teku/BeaconNode.java
@@ -15,24 +15,18 @@ package tech.pegasys.teku;
 
 import static tech.pegasys.teku.logging.StatusLogger.STATUS_LOG;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.eventbus.AsyncEventBus;
 import com.google.common.eventbus.EventBus;
-import com.google.common.eventbus.SubscriberExceptionContext;
-import com.google.common.eventbus.SubscriberExceptionHandler;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import io.vertx.core.Vertx;
-import java.lang.reflect.Method;
 import java.nio.file.Path;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import org.hyperledger.besu.plugin.services.MetricsSystem;
 import tech.pegasys.teku.data.recorder.SSZTransitionRecorder;
-import tech.pegasys.teku.events.ChannelExceptionHandler;
 import tech.pegasys.teku.events.EventChannels;
 import tech.pegasys.teku.logging.LoggingConfiguration;
 import tech.pegasys.teku.logging.LoggingConfigurator;
-import tech.pegasys.teku.logging.StatusLogger;
 import tech.pegasys.teku.metrics.MetricsEndpoint;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.services.ServiceController;
@@ -57,8 +51,8 @@ public class BeaconNode {
 
     this.metricsEndpoint = new MetricsEndpoint(config, vertx);
     final MetricsSystem metricsSystem = metricsEndpoint.getMetricsSystem();
-    final EventBusExceptionHandler subscriberExceptionHandler =
-        new EventBusExceptionHandler(STATUS_LOG);
+    final TekuDefaultExceptionHandler subscriberExceptionHandler =
+        new TekuDefaultExceptionHandler();
     this.eventChannels = new EventChannels(subscriberExceptionHandler, metricsSystem);
     this.eventBus = new AsyncEventBus(threadPool, subscriberExceptionHandler);
 
@@ -97,60 +91,5 @@ public class BeaconNode {
     eventChannels.stop();
     metricsEndpoint.stop();
     vertx.close();
-  }
-}
-
-@VisibleForTesting
-final class EventBusExceptionHandler
-    implements SubscriberExceptionHandler, ChannelExceptionHandler {
-
-  private final StatusLogger log;
-
-  EventBusExceptionHandler(final StatusLogger log) {
-    this.log = log;
-  }
-
-  @Override
-  public void handleException(final Throwable exception, final SubscriberExceptionContext context) {
-    handleException(
-        exception,
-        "event '"
-            + context.getEvent().getClass().getName()
-            + "'"
-            + " in handler '"
-            + context.getSubscriber().getClass().getName()
-            + "'"
-            + " (method  '"
-            + context.getSubscriberMethod().getName()
-            + "')");
-  }
-
-  @Override
-  public void handleException(
-      final Throwable error,
-      final Object subscriber,
-      final Method invokedMethod,
-      final Object[] args) {
-    handleException(
-        error,
-        "event '"
-            + invokedMethod.getDeclaringClass()
-            + "."
-            + invokedMethod.getName()
-            + "' in handler '"
-            + subscriber.getClass().getName()
-            + "'");
-  }
-
-  private void handleException(final Throwable exception, final String subscriberDescription) {
-    if (isSpecFailure(exception)) {
-      log.specificationFailure(subscriberDescription, exception);
-    } else {
-      log.unexpectedFailure(subscriberDescription, exception);
-    }
-  }
-
-  private static boolean isSpecFailure(final Throwable exception) {
-    return exception instanceof IllegalArgumentException;
   }
 }
